+++ serde_arrow/src/internal/schema.rs
@@ -1,3 +1,5 @@
+mod deserialization;
+
 use std::{
     collections::{BTreeMap, HashMap},
     str::FromStr,
@@ -257,56 +259,6 @@ pub trait SchemaLike: Sized + Sealed {
 #[derive(Default, Debug, PartialEq, Clone, Serialize)]
 pub struct SerdeArrowSchema {
     pub(crate) fields: Vec<GenericField>,
-}
-
-impl<'de> serde::Deserialize<'de> for SerdeArrowSchema {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        struct Visitor;
-
-        impl<'de> serde::de::Visitor<'de> for Visitor {
-            type Value = SerdeArrowSchema;
-
-            fn expecting(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-                write!(f, "a sequence of fields or a struct with key 'fields' containing a sequence of fields")
-            }
-
-            fn visit_seq<A: serde::de::SeqAccess<'de>>(
-                self,
-                mut seq: A,
-            ) -> Result<Self::Value, A::Error> {
-                let mut fields = Vec::new();
-
-                while let Some(item) = seq.next_element::<GenericField>()? {
-                    fields.push(item);
-                }
-
-                Ok(SerdeArrowSchema { fields })
-            }
-
-            fn visit_map<A: serde::de::MapAccess<'de>>(
-                self,
-                mut map: A,
-            ) -> Result<Self::Value, A::Error> {
-                use serde::de::Error;
-
-                let mut fields = None;
-
-                while let Some(key) = map.next_key::<String>()? {
-                    if key == "fields" {
-                        fields = Some(map.next_value::<Vec<GenericField>>()?);
-                    }
-                }
-
-                let Some(fields) = fields else {
-                    return Err(A::Error::custom("missing field `fields`"));
-                };
-
-                Ok(SerdeArrowSchema { fields })
-            }
-        }
-
-        deserializer.deserialize_any(Visitor)
-    }
 }
 
 impl SerdeArrowSchema {
@@ -637,7 +589,7 @@ impl std::str::FromStr for GenericDataType {
 
             Ok(GenericDataType::Decimal128(precision, scale))
         } else {
-            fail!("cannot parse data type")
+            fail!("cannot parse data type {s}");
         }
     }
 }