+++ Changes.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 0.10.1
+
+- Allow to use `arrow` schemas in `SchemaLike::from_value()`, e.g., `let fields
+  = Vec::<Field>::from_value(&batch.schema())`.
+
 ## 0.10.0
 
 - Remove deprecated APIs